# web/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strategon Control Plane",
    "version": "1.0.0",
    "description": "Human API for the Strategon control plane (Connect JSON over HTTP).\n\nAuthenticate with a session cookie (UI login) or a Bearer API token\n(`Authorization: Bearer str_live_…`) minted from the UI **API tokens** page.\n\nAgent gRPC/Connect endpoints are served on a separate port and are not\nincluded in this document.\n"
  },
  "servers": [
    {
      "url": "/",
      "description": "Human API (same origin when using the embedded UI)"
    }
  ],
  "paths": {
...
    "/strategyplatform.v1.ControlPlaneService/PutSecret": {
      "post": {
        "tags": [
          "ControlPlaneService"
        ],
        "summary": "PutSecret",
        "operationId": "strategyplatform.v1.ControlPlaneService.PutSecret",
        "parameters": [
          {
            "name": "Connect-Protocol-Version",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/connect-protocol-version"
            }
          },
          {
            "name": "Connect-Timeout-Ms",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/connect-timeout-header"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/strategyplatform.v1.PutSecretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connect.error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/strategyplatform.v1.PutSecretResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "connect-protocol-version": {
        "type": "number",
        "title": "Connect-Protocol-Version",
        "enum": [
          1
        ],
        "description": "Define the version of the Connect protocol",
        "const": 1
      },
      "connect-timeout-header": {
        "type": "number",
        "title": "Connect-Timeout-Ms",
        "description": "Define the timeout, in ms"
      },
      "connect.error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "examples": [
              "not_found"
            ],
            "enum": [
              "canceled",
              "unknown",
              "invalid_argument",
              "deadline_exceeded",
              "not_found",
              "already_exists",
              "permission_denied",
              "resource_exhausted",
              "failed_precondition",
              "aborted",
              "out_of_range",
              "unimplemented",
              "internal",
              "unavailable",
              "data_loss",
              "unauthenticated"
            ],
            "description": "The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code]."
          },
          "message": {
            "type": "string",
            "description": "A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client."
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/connect.error_details.Any"
            },
            "description": "A list of messages that carry the error details. There is no limit on the number of messages."
          }
        },
        "title": "Connect Error",
        "additionalProperties": true,
        "description": "Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation"
      },
      "connect.error_details.Any": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URL that acts as a globally unique identifier for the type of the serialized message. For example: `type.googleapis.com/google.rpc.ErrorInfo`. This is used to determine the schema of the data in the `value` field and is the discriminator for the `debug` field."
          },
          "value": {
            "type": "string",
            "format": "binary",
            "description": "The Protobuf message, serialized as bytes and base64-encoded. The specific message type is identified by the `type` field."
          },
          "debug": {
            "oneOf": [
              {
                "type": "object",
                "title": "Any",
                "additionalProperties": true,
                "description": "Detailed error information."
              }
            ],
            "discriminator": {
              "propertyName": "type"
            },
            "title": "Debug",
            "description": "Deserialized error detail payload. The 'type' field indicates the schema. This field is for easier debugging and should not be relied upon for application logic."
          }
        },
        "additionalProperties": true,
        "description": "Contains an arbitrary serialized message along with a @type that describes the type of the serialized message, with an additional debug field for ConnectRPC error details."
      },
...
      "strategyplatform.v1.PutSecretRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "title": "name",
            "description": "e.g. \"db-url\" (not \"secret.db-url\")"
          },
          "value": {
            "type": "string",
            "title": "value",
            "description": "plaintext in; size-capped server-side"
          }
        },
        "title": "PutSecretRequest",
        "additionalProperties": false,
        "description": "Secret catalog: ciphertext at rest; human APIs never return plaintext.\n Public token on the wire is \"secret.\u003cname\u003e\" (PutSecret, env, member.vars)."
      },
      "strategyplatform.v1.PutSecretResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "title": "token",
            "description": "\"secret.\u003cname\u003e\""
          }
        },
        "title": "PutSecretResponse",
        "additionalProperties": false
      },
...
    }
  },
  "security": [],
...
}